[add]: LoginViewModel 파일 추가 #61
